requirements.txt
pymongo
dotenv
google-genai
ollama